requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.3.9
cssselect==1.1.0
decorator==5.1.0
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.4.0
Flask==2.0.2
huggingface-hub==0.2.1
idna==3.3
itsdangerous==2.0.1
jieba3k==0.35.1
Jinja2==3.0.3
joblib==1.1.0
lxml==4.6.4
MarkupSafe==2.0.1
newspaper3k==0.2.8
nltk==3.6.5
numpy==1.21.4
packaging==21.3
pandas==1.3.4
Pillow==8.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
requests-file==1.5.1
rouge==1.0.1
sacremoses==0.0.46
sentence-transformers==2.1.0
sentencepiece==0.1.96
scikit-learn==1.0.1
scipy==1.7.3
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.1
threadpoolctl==3.0.0
tinysegmenter==0.3
tldextract==3.1.2
tokenizers==0.10.3
torchaudio==0.10.0+cu111
torchsummary==1.5.1
torchtext==0.11.0
torchvision==0.11.1+cu111
tqdm==4.62.3
transformers==4.12.5
typing-extensions==4.0.1
urllib3==1.26.7
validators==0.18.2
Werkzeug==2.0.2
wordnet==0.0.1b2
